+++ padogrid-deployment/src/main/resources/common/bin_sh/.argenv.sh
@@ -472,7 +472,7 @@ fi
 # SED backup prefix
 if [[ ${OS_NAME} == DARWIN* ]]; then
    # Mac - space required
-   __SED_BACKUP=" 0"
+   __SED_BACKUP=" ''"
 else
    __SED_BACKUP="0"
 fi