+++ padogrid-deployment/src/main/resources/common/bin_sh/.utilenv.sh
@@ -2598,7 +2598,7 @@ cat <<EOF
 |   ___/  /  /_\  \   |  |  |  |  |  |  | |  | |_ | |      /     |  | |  |  |  |
 |  |     /  _____  \  |  '--'  |  '--'  | |  |__| | |  |\  \----.|  | |  '--'  |
 | _|    /__/     \__\ |_______/ \______/   \______| | _| '._____||__| |_______/ 
-Copyright 2020 Netcrest Technologies, LLC. All rights reserved.
+Copyright 2020,2021 Netcrest Technologies, LLC. All rights reserved.
 v$PADOGRID_VERSION
 EOF
 echo -e "Manual: ${CUrl}https://github.com/padogrid/padogrid/wiki${CNone}"